requirements.txt
fastapi
uvicorn
pydantic
google-genai
requests
python-dotenv
streamlit
huggingface_hub
elevenlabs
moviepy